Mask and ordering fixes

+++ model/modules/masked.py
@@ -10,14 +10,14 @@ MaskMode = Enum('MaskMode', 'subtract multiply')
 MaskTime = Enum('MaskTime', 'pre post')
 
 
-def mask_sequence(input_batch: t.Tensor, mask_index: float=0) -> t.LongTensor:
+def mask_sequence(input_batch: t.Tensor, mask_index: float=0) -> t.ByteTensor:
     """
     Returns a LongTensor where masked indices are 0 and rest 1
     :param input_batch: Batch first tensor of padded sequences
     :param mask_index: Value that signifies an index that should be masked
-    :returns: A LongTensor, same shape as input_batch
+    :returns: A ByteTensor, same shape as input_batch
     """
-    return t.LongTensor((input_batch != mask_index).long())
+    return t.ByteTensor((input_batch != mask_index))
 
 
 class MaskedOp(nn.Module):
@@ -40,7 +40,7 @@ class MaskedOp(nn.Module):
         self.mask_time = mask_time
         self.mask_value = mask_value
 
-    def _apply_mask(self, inpt: t.Tensor, mask: t.LongTensor) -> Tensor:
+    def _apply_mask(self, inpt: t.Tensor, mask: t.ByteTensor) -> Tensor:
         """
         Applies the mask to inpt according to MaskMode
         :param inpt: Input tensor to apply mask to
@@ -55,7 +55,7 @@ class MaskedOp(nn.Module):
         else:
             raise Exception('Malformed mask_mode for MaskedOp: %s' % self.mask_mode)
 
-    def forward(self, input_batch: t.Tensor, mask: t.LongTensor) -> t.Tensor:
+    def forward(self, input_batch: t.Tensor, mask: t.ByteTensor) -> t.Tensor:
         """
         Subtracts mask from its input and applies op to it
         :param input_batch: Batch of variable length sequences to mask and apply op on
